+++ plugin/src/ui/query/displays/EmptyDisplay.tsx
@@ -11,5 +11,7 @@ export const EmptyDisplay: React.FC<Props> = ({ message }) => {
   const i18n = t().query.displays.empty;
   const displayMessage = message ?? i18n.label;
 
-  return <Callout className="todoist-no-tasks" title={displayMessage} iconId="info" />;
+  return (
+    <Callout className="todoist-no-tasks" title={displayMessage} iconId="check" variant="success" />
+  );
 };